Allow floating-point thread, memory reservations.
A floating point thread reservation would make sense in a context where
a stage has a thread that spends some time sleeping, e.g. polling for a
file to show up.

I local mode, the job manager now schedules jobs at a granularity of 1MB
for memory or 1% of a thread for cores.  The total reservable values
specified by `--localcores` or `--localmem` are still integers, however.

While in the neighborhood, also allow specifying `volatile = false` in a
stage definition, in preparation for a future change making `volatile =
strict` the default.

Also, AST nodes now keep track of their column as well as line numbers.

+++ martian/core/jobinfo.go
@@ -15,9 +15,9 @@ type JobInfo struct {
 	MemoryUsage   *ObservedMemory   `json:"used_bytes,omitempty"`
 	IoStats       *IoStats          `json:"io,omitempty"`
 	WallClockInfo *WallClockInfo    `json:"wallclock,omitempty"`
-	Threads       int               `json:"threads,omitempty"`
-	MemGB         int               `json:"memGB,omitempty"`
-	VMemGB        int               `json:"vmemGB,omitempty"`
+	Threads       float64           `json:"threads,omitempty"`
+	MemGB         float64           `json:"memGB,omitempty"`
+	VMemGB        float64           `json:"vmemGB,omitempty"`
 	ProfileConfig *ProfileConfig    `json:"profile_config,omitempty"`
 	ProfileMode   ProfileMode       `json:"profile_mode,omitempty"`
 	Stackvars     string            `json:"stackvars_flag,omitempty"`
